Refaktorer - trekk ut metode, fjern error i konsoll
