+++ src/søknad/steg/4-barnasbosted/bostedOgSamvær/BorForelderINorge.tsx
@@ -1,6 +1,10 @@
 import React, { FC } from 'react';
-import { IForelder } from '../../../../models/steg/forelder';
-import { ISpørsmål, ISvar } from '../../../../models/felles/spørsmålogsvar';
+import { EForelder, IForelder } from '../../../../models/steg/forelder';
+import {
+  ESvar,
+  ISpørsmål,
+  ISvar,
+} from '../../../../models/felles/spørsmålogsvar';
 import KomponentGruppe from '../../../../components/gruppe/KomponentGruppe';
 import { borINorge } from '../ForeldreConfig';
 import { Input } from 'nav-frontend-skjema';
@@ -9,21 +13,55 @@ import { useIntl } from 'react-intl';
 import JaNeiSpørsmålMedNavn from '../../../../components/spørsmål/JaNeiSpørsmålMedNavn';
 import { IBarn } from '../../../../models/steg/barn';
 import { hentBarnNavnEllerBarnet } from '../../../../utils/barn';
+import {
+  erJaNeiSvar,
+  hentBooleanFraValgtSvar,
+} from '../../../../utils/spørsmålogsvar';
 
 interface Props {
   barn: IBarn;
-  settFelt: (spørsmål: ISpørsmål, svar: ISvar) => void;
   forelder: IForelder;
   settForelder: (verdi: IForelder) => void;
+  settDokumentasjonsbehovForBarn: (
+    spørsmål: ISpørsmål,
+    valgtSvar: ISvar,
+    barneid: string,
+    barnapassid?: string
+  ) => void;
 }
 
 const BorForelderINorge: FC<Props> = ({
   settForelder,
   barn,
   forelder,
-  settFelt,
+  settDokumentasjonsbehovForBarn,
 }) => {
   const intl = useIntl();
+
+  const settFelt = (spørsmål: ISpørsmål, svar: ISvar) => {
+    const nyForelder = {
+      ...forelder,
+      [spørsmål.søknadid]: {
+        spørsmålid: spørsmål.søknadid,
+        svarid: svar.id,
+        label: hentTekst(spørsmål.tekstid, intl),
+        verdi: erJaNeiSvar(svar)
+          ? hentBooleanFraValgtSvar(svar)
+          : svar.svar_tekst,
+      },
+    };
+
+    if (
+      spørsmål.søknadid === EForelder.borINorge &&
+      nyForelder.land &&
+      svar.id === ESvar.JA
+    ) {
+      delete nyForelder.land;
+    }
+    settForelder(nyForelder);
+    settDokumentasjonsbehovForBarn(spørsmål, svar, barn.id);
+  };
+
   return (
     <>
       <KomponentGruppe>
